+++ src/PackageInfo.php
@@ -1,8 +1,0 @@
-<?php
-namespace Icecave\Parity;
-
-class PackageInfo
-{
-    const NAME = 'Parity';
-    const VERSION = '1.0.0';
-}