+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 3.0.0 (2020-08-25)
 
 - **[BC]** Drop support for PHP 7.1
+- **[BC]** Remove `PackageInfo` class
 
 ### 2.0.0 (2018-11-06)